webhook_service/requirements.txt
flask>=3.0.0
PyGithub>=2.1.0
google-cloud-aiplatform>=1.118.0
google-cloud-secret-manager>=2.16.0
google-cloud-firestore>=2.14.0
cryptography>=41.0.0
gunicorn>=21.2.0
pyyaml>=6.0.0
pytest>=8.3.4
pytest-mock>=3.14.0
